Revert "Merge branch 'master' into '20200210-sj-33'"
This reverts commit 7c2bdd272a763cc85046dc45e9afd95bd8eeb7a0, reversing
changes made to 24d41cca12d979f63b4b1544aecffd8e78cd8743.

+++ src/main/java/com/kh/hotels/mngApproval/model/service/ApprovalService.java
@@ -76,7 +76,7 @@ public interface ApprovalService {
 
 	ArrayList<HashMap<String, Object>> selectOrderDocu();
 
-	ArrayList<HashMap<String, Object>> selectOrderList(Long dnum);
+	ArrayList<HashMap<String, Object>> selectOrderList(int dnum);
 
 	int insertOrderList(ArrayList<OrderRequest> oRequestList);
